config, reformatting
