+++ ALKIS_WFS_DOWNLOAD.calc_sfl_bodenschaetzung.pyt.xml
@@ -1,0 +1,2 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<metadata xml:lang="de"><Esri><CreaDate>20260108</CreaDate><CreaTime>09363100</CreaTime><ArcGISFormat>1.0</ArcGISFormat><SyncOnce>TRUE</SyncOnce></Esri></metadata>
