+++ ALKIS_WFS_DOWNLOAD.pyt.xml
@@ -1,2 +1,2 @@
 <?xml version="1.0"?>
-<metadata xml:lang="de"><Esri><CreaDate>20250401</CreaDate><CreaTime>10185200</CreaTime><ArcGISFormat>1.0</ArcGISFormat><SyncOnce>TRUE</SyncOnce><ModDate>20260107</ModDate><ModTime>155114</ModTime><scaleRange><minScale>150000000</minScale><maxScale>5000</maxScale></scaleRange><ArcGISProfile>ItemDescription</ArcGISProfile></Esri><toolbox name="ALKIS_WFS_DOWNLOAD" alias="ALKISWFSDownload"><arcToolboxHelpPath>c:\program files\arcgis\pro\Resources\Help\gp</arcToolboxHelpPath><toolsets/></toolbox><dataIdInfo><idCitation><resTitle>ALKIS_WFS_DOWNLOAD</resTitle></idCitation><idPurp>ALKIS_WFS_DOWNLOAD</idPurp><searchKeys><keyword>ALKIS_WFS_DOWNLOAD</keyword></searchKeys></dataIdInfo><distInfo><distributor><distorFormat><formatName>ArcToolbox Toolbox</formatName></distorFormat></distributor></distInfo><mdHrLv><ScopeCd value="005"/></mdHrLv><mdDateSt Sync="TRUE">20250407</mdDateSt></metadata>
+<metadata xml:lang="de"><Esri><CreaDate>20250401</CreaDate><CreaTime>10185200</CreaTime><ArcGISFormat>1.0</ArcGISFormat><SyncOnce>TRUE</SyncOnce><ModDate>20260114</ModDate><ModTime>081311</ModTime><scaleRange><minScale>150000000</minScale><maxScale>5000</maxScale></scaleRange><ArcGISProfile>ItemDescription</ArcGISProfile></Esri><toolbox name="ALKIS_WFS_DOWNLOAD" alias="ALKISWFSDownload"><arcToolboxHelpPath>c:\program files\arcgis\pro\Resources\Help\gp</arcToolboxHelpPath><toolsets/></toolbox><dataIdInfo><idCitation><resTitle>ALKIS_WFS_DOWNLOAD</resTitle></idCitation><idPurp>ALKIS_WFS_DOWNLOAD</idPurp><searchKeys><keyword>ALKIS_WFS_DOWNLOAD</keyword></searchKeys></dataIdInfo><distInfo><distributor><distorFormat><formatName>ArcToolbox Toolbox</formatName></distorFormat></distributor></distInfo><mdHrLv><ScopeCd value="005"/></mdHrLv><mdDateSt Sync="TRUE">20250407</mdDateSt></metadata>
