+++ sfl/calc_sfl_nutzung.py
@@ -11,17 +11,9 @@ import arcpy
 import os
 import time
 import pandas as pd
-import numpy as np
 from sfl.init_dataframes import DataFrameLoader
 from config.config_loader import FieldConfigLoader
-
-try:
-    from shapely.geometry import shape
-
-    SHAPELY_AVAILABLE = True
-except ImportError:
-    SHAPELY_AVAILABLE = False
-    arcpy.AddWarning("Shapely nicht verfügbar - Mini-Flächen-Merge wird übersprungen")
+from sfl.merge_mini_geometries import merge_mini_geometries
 
 
 class SFLCalculatorNutzung(DataFrameLoader):
@@ -29,7 +21,9 @@ class SFLCalculatorNutzung(DataFrameLoader):
     Optimierte Klasse für SFL- und EMZ-Berechnungen mit Pandas/NumPy Vectorisierung.
     """
 
-    def __init__(self, gdb_path, workspace, keep_workdata, flaechenformindex, max_shred_area, merge_area):
+    def __init__(
+        self, gdb_path, workspace, keep_workdata, flaechenformindex, max_shred_area, merge_area, delete_unmerged_mini
+    ):
         """
         Initialisiert den optimierten SFL Calculator.
 
@@ -43,6 +37,7 @@ class SFLCalculatorNutzung(DataFrameLoader):
         )
 
         self.merge_area = merge_area  # Splitterflächengröße für Kleinstflächen, die ohne Merge in angrenzende Geometrie erhalten bleiben
+        self.delete_unmerged_mini = delete_unmerged_mini
 
         self.flaechenformindex = flaechenformindex
 
@@ -56,7 +51,9 @@ class SFLCalculatorNutzung(DataFrameLoader):
         self.geom_cache_nutzung = {}  # Geometry Cache für Nutzung Features
 
         self.keep_workdata = keep_workdata
+
         self.cfg = FieldConfigLoader.load_config()
+
         self.nutz = self.cfg["nutzung"]
         self.flst = self.cfg["flurstueck"]
 
@@ -124,7 +121,7 @@ class SFLCalculatorNutzung(DataFrameLoader):
             # Laden in DataFrame
             if not self.load_flurstuecke_to_dataframe():
                 return False
-            if not self.load_nutzung_to_dataframe():
+            if not self.load_nutzung_to_dataframe("nutzung_dissolve"):
                 return False
 
             # Merge DataFrames auf FSK um Verbesserungsfaktor zu bekommen
@@ -139,46 +136,13 @@ class SFLCalculatorNutzung(DataFrameLoader):
             df["raw_sfl"] = df["geom_area"] * df["verbesserung"]
             df["sfl"] = (df["raw_sfl"] + 0.5).astype(int)  # round-half-up
 
-            # Kleinstflächen-Filterung pro FSK
-            mask_mini = (df["sfl"] <= self.max_shred_qm) & (df["amtliche_flaeche"] > self.max_shred_qm)
-
-            df.loc[mask_mini, "is_mini"] = True
-            df.loc[~mask_mini, "is_mini"] = False
-
-            # Separate mini und non-mini Features
-            df_mini = df[df["is_mini"] == True].copy()
-            df_main = df[df["is_mini"] == False].copy()
-
-            arcpy.AddMessage(f"  Identifiziert {len(df_mini)} Kleinstflächen zur Verarbeitung")
-
-            # Mini-Flächen-Filterung: Nur die mergen, die WENIGER als 1 m² bei Verteilung ergeben
-            if len(df_mini) > 0:
-                # Trennung: erhaltungswürdig (>= 1) vs. zu mergen (< 1)
-                mask_keep = df_mini["sfl"] >= self.merge_area
-                df_mini["perimeter"] = df_mini["geometry"].apply(lambda geom: geom.length)
-                df_mini["form_index"] = df_mini["perimeter"] / np.sqrt(df_mini["geom_area"])
-
-                # Schmale, lange Schnipsel filtern (form_index < flaechenformindex_input = sehr dünn)
-                mask_real_feature = df_mini["form_index"] < self.flaechenformindex
-                df_mini_keep = df_mini[(mask_keep) & (mask_real_feature)].copy()
-                df_mini_merge = df_mini[(~mask_keep) | (~mask_real_feature)].copy()
-
-                arcpy.AddMessage(
-                    f"    {len(df_mini_keep)} Mini-Flächen erhalten (>= {self.merge_area} m² und Flächenformindex <{self.flaechenformindex}), "
-                    f"{len(df_mini_merge)} Mini-Flächen werden gemergt (< {self.merge_area} m²) oder Flächenformindex >={self.flaechenformindex})"
-                )
-
-                # Erhaltungswürdige Mini-Flächen zu Main hinzufügen
-                df_main = pd.concat([df_main, df_mini_keep], ignore_index=True)
-
-                # Merge zu verlustende Mini-Flächen mit angrenzenden Hauptflächen
-                # Nach dem Merge: ungemergte werden zu Main hinzugefügt, gemergte werden aus df_mini entfernt
-                if len(df_mini_merge) > 0 and SHAPELY_AVAILABLE:
-                    df_main_after_merge, df_mini_to_delete = self._merge_mini_into_main_nutzung(df_main, df_mini_merge)
-                    df_main = df_main_after_merge
-                    df_mini = df_mini_to_delete  # Nur die tatsächlich gemergt wurden
-                else:
-                    df_mini = pd.DataFrame()  # Nichts zu mergen
+            df_main, df_mini, df_not_merged = merge_mini_geometries(
+                df, self.max_shred_qm, self.merge_area, self.flaechenformindex
+            )
+            if self.delete_unmerged_mini:
+                df_mini = pd.concat([df_mini, df_not_merged], ignore_index=True)
+            else:
+                df_main = pd.concat([df_main, df_not_merged], ignore_index=True)
 
             # Overlap-Handling: weitere_nutzung_id == 1000
             overlap_mask = df_main["weitere_nutzung_id"] == 1000
@@ -197,122 +161,6 @@ class SFLCalculatorNutzung(DataFrameLoader):
         except Exception as e:
             arcpy.AddError(f"Fehler bei vectorized_calculate_sfl_nutzung: {str(e)}")
             return False
-
-    def _merge_mini_into_main_nutzung(self, df_main, df_mini):
-        """
-        Merge Mini-Flächen mit angrenzenden Hauptflächen
-        - Fallback-Strategien: touches → intersects → buffer-distance
-        """
-        arcpy.AddMessage("  Merge Kleinstflächen mit Hauptflächen...")
-
-        start_time = time.time()
-        merged_oids = set()
-        tolerance = 0.1  # 10cm Buffer für Toleranz
-        total_mini = len(df_mini)
-        processed_mini = 0
-
-        # Loop durch Mini-Flächen (äußerer Loop = klein!)
-        for _, mini_row in df_mini.iterrows():
-            processed_mini += 1
-            mini_oid = mini_row["objectid"]
-            mini_geom = mini_row["geometry"]
-            mini_fsk = mini_row["fsk"]
-
-            # Progress alle 100 Features oder am Ende
-            if processed_mini % 1000 == 0 or processed_mini == total_mini:
-                elapsed = time.time() - start_time
-                arcpy.AddMessage(
-                    f"    Fortschritt: {processed_mini}/{total_mini} Mini-Flächen verarbeitet "
-                    f"({len(merged_oids)} erfolgreich gemergt, {elapsed:.1f}s)"
-                )
-
-            # Hole nur Main-Features dieser FSK
-            fsk_main_mask = df_main["fsk"] == mini_fsk
-            if not fsk_main_mask.any():
-                continue
-
-            fsk_main_idx = df_main[fsk_main_mask].index
-            best_match_idx = None
-
-            # === Strategie 1: Direct touches/intersects ===
-            for main_idx in fsk_main_idx:
-                main_geom = df_main.at[main_idx, "geometry"]
-                try:
-                    if main_geom.touches(mini_geom) or main_geom.intersects(mini_geom):
-                        best_match_idx = main_idx
-                        break
-                except:
-                    pass
-
-            # === Strategie 2: Distance check mit Toleranz ===
-            if best_match_idx is None:
-                min_distance = float("inf")
-                for main_idx in fsk_main_idx:
-                    main_geom = df_main.at[main_idx, "geometry"]
-                    try:
-                        distance = main_geom.distance(mini_geom)
-                        if distance < tolerance and distance < min_distance:
-                            min_distance = distance
-                            best_match_idx = main_idx
-                    except:
-                        pass
-
-            # === Strategie 3: Buffer - größte angrenzende Fläche ===
-            if best_match_idx is None:
-                try:
-                    mini_buffer = mini_geom.buffer(0.5)
-                    max_area = 0
-                    for main_idx in fsk_main_idx:
-                        main_geom = df_main.at[main_idx, "geometry"]
-                        try:
-                            if mini_buffer.intersects(main_geom):
-                                if main_geom.area > max_area:
-                                    max_area = main_geom.area
-                                    best_match_idx = main_idx
-                        except:
-                            pass
-                except:
-                    pass
-
-            # === Merge durchführen ===
-            if best_match_idx is not None:
-                try:
-                    main_geom = df_main.at[best_match_idx, "geometry"]
-                    union_geom = main_geom.union(mini_geom)
-                    union_area = union_geom.area
-
-                    df_main.at[best_match_idx, "geometry"] = union_geom
-                    df_main.at[best_match_idx, "geom_area"] = union_area
-
-                    # SFL mit neuer Fläche berechnen
-                    verbesserung = df_main.at[best_match_idx, "verbesserung"]
-                    new_sfl = int(union_area * verbesserung + 0.5)
-                    df_main.at[best_match_idx, "sfl"] = new_sfl
-
-                    merged_oids.add(mini_oid)
-                except Exception as e:
-                    arcpy.AddWarning(f"    Merge für Mini {mini_oid} fehlgeschlagen: {e}")
-
-        df_mini_deleted = df_mini[df_mini["objectid"].isin(merged_oids)].copy()
-        df_mini_not_merged = df_mini[~df_mini["objectid"].isin(merged_oids)].copy()
-        elapsed = time.time() - start_time
-
-        # Warnung für nicht-gemergte Flächen
-        if len(df_mini_not_merged) > 0:
-            arcpy.AddWarning(
-                f"    WARNUNG: {len(df_mini_not_merged)} Mini-Flächen konnten nicht an angrenzendes Flurstück angeschmiegt werden und werden gelöscht"
-            )
-            for _, row in df_mini_not_merged.iterrows():
-                fsk = row["fsk"]
-                oid = row["objectid"]
-                sfl = row["sfl"]
-                arcpy.AddWarning(f"      FSK {fsk}, OID {oid}, SFL {sfl} m² - GELÖSCHT")
-
-        arcpy.AddMessage(f"    {len(merged_oids)}/{len(df_mini)} Mini-Flächen gemergt ({elapsed:.2f}s)")
-
-        # Alle Mini-Flächen zurückgeben (gemergt + nicht-gemergt) zum Löschen
-        all_mini_to_delete = pd.concat([df_mini_deleted, df_mini_not_merged], ignore_index=True)
-        return df_main, all_mini_to_delete
 
     def _apply_delta_correction_nutzung(self, df):
         """Delta-Korrektur je FSK: kleine Deltas (<5 qm) proportional auf große Features verteilen."""
@@ -501,11 +349,15 @@ class SFLCalculatorNutzung(DataFrameLoader):
             return False
 
 
-def calculate_sfl_nutzung(gdb_path, workspace, keep_workdata, flaechenformindex, max_shred_area, merge_area):
+def calculate_sfl_nutzung(
+    gdb_path, workspace, keep_workdata, flaechenformindex, max_shred_area, merge_area, delete_unmerged_mini
+):
     """
     :return: True bei Erfolg, False bei Fehler
     """
-    calculator = SFLCalculatorNutzung(gdb_path, workspace, keep_workdata, flaechenformindex, max_shred_area, merge_area)
+    calculator = SFLCalculatorNutzung(
+        gdb_path, workspace, keep_workdata, flaechenformindex, max_shred_area, merge_area, delete_unmerged_mini
+    )
 
     # if not calculator.prepare_nutzung():
     #     return False
